requirements.txt
langchain
chainlit
langchain_openai
openai
chromadb
tiktoken
pymupdf
langchain_community
chainlit
langchain
python-dotenv